+++ app/routes/components/Avatar/changelog.codegen.ts
@@ -5,4 +5,5 @@ export const changelog = ''
   + '- 2025-09-23: Initial version.\n'
   + '- 2025-10-03: Add `Avatar.Img` for image support. Update styles to ensure images\n'
   + '  fit within the avatar circle.\n'
+  + '- 2025-10-11: Using type `ColorName` for color prop.\n'
   ;